fix: move storageAware functions to separate file storage.go
